chore: refactor project files structure
