+++ src/pages/index.ts
@@ -1,6 +1,6 @@
 export { default as PageAbout } from "./about/about.page.svelte";
-export { default as PageIllumination } from "./illumination/illumination.page.svelte";
+export { default as PageLights } from "./lights/lights.page.svelte";
 export { default as PageMoon } from "./moon/moon.page.svelte";
 export { default as PageSun } from "./sun/sun.page.svelte";
 export { default as PageTimeline } from "./timeline/timeline.page.svelte";
-export { default as PageToday } from "./illumination/today.svelte";
+export { default as PageToday } from "./lights/lights-today.svelte";
