+++ src/pages/lights/lights.page.svelte
