+++ src/pages/timeline/timeline.page.svelte
@@ -1,9 +1,9 @@
 <script lang="ts">
-	import { Timeline, TimelineSection, TimelineEntry } from "../../components";
-	import { getTimeline } from "../../services/timeline";
-	import { incrementDateByDay } from "../../helpers";
+	import { Timeline, TimelineSection, TimelineEntry } from "@lib/components";
+	import { getTimeline } from "@services/timeline";
+	import { incrementDateByDay } from "@lib/helpers";
 	import Icon from "./timeline.icon.svelte";
-	import { dict } from "../../dict";
+	import { dict } from "@lib/dict";
 
 	interface Timeline {
 		date: Date;
